Refactor event service structure and update package imports
- Changed package imports from `services` to `service` for consistency across EmitEventController, EventSummaryController, and EventTypeController.
- Deleted old service implementation classes: EmitEventServiceImpl, EventSummaryServiceImpl, and EventTypeServiceImpl.
- Introduced new service interfaces: EmitEventService, EventSummaryService, and EventTypeService.
- Implemented new service classes: EmitEventServiceImpl, EventSummaryServiceImpl, and EventTypeServiceImpl with appropriate logic.
- Updated tests for EventSummaryServiceImpl to reflect new service structure and ensure functionality.
- Enhanced ApiErrorAuthenticationEntryPoint and ToolPriorityTuningService to utilize Clock for time management.
