+++ pos-event-receiver/src/test/java/com/positivity/poseventreceiver/internal/service/EventSummaryServiceImplTest.java
@@ -1,4 +1,4 @@
-package com.positivity.poseventreceiver.internal.services;
+package com.positivity.poseventreceiver.internal.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
